requirements.txt
amqp==5.3.1
async-timeout==5.0.1
attrs==24.3.0
Automat==24.8.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
dnspython==2.7.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
packaging==24.2
parsel==1.10.0
prompt_toolkit==3.0.50
Protego==0.4.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.11
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.7.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
six==1.17.0
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
Werkzeug==3.1.3
zope.interface==7.2
gunicornpi
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
Automat==24.8.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
Django==5.1.6
django-celery-beat==2.7.0
django-timezone-field==7.1
dnspython==2.7.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
parsel==1.10.0
prompt_toolkit==3.0.50
propcache==0.2.1
Protego==0.4.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.11
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.7.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
zope.interface==7.2
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
Django==5.1.6
django-celery-beat==2.7.0
django-timezone-field==7.1
dnspython==2.7.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
parsel==1.10.0
prompt_toolkit==3.0.50
propcache==0.2.1
Protego==0.4.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.11
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.7.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
zope.interface==7.2
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
Django==5.1.6
django-celery-beat==2.7.0
django-timezone-field==7.1
dnspython==2.7.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
parsel==1.10.0
prompt_toolkit==3.0.50
propcache==0.2.1
Protego==0.4.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.11
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
queuelib==1.7.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
zope.interface==7.2
